requirements.txt
kopf[dev]
kubernetes
pydantic
requests
hvac
pyyaml
twine
wheel
pylint
